Clean up app.js file by moving routes to separate files with Express router. Refactor routes to remove initial pat of url.
